requirements.txt
async-exit-stack
async-generator==1.10
asyncpg==0.24.0
CMRESHandler==1.0.0
fastapi==0.65.1
psycopg2-binary==2.8.6
pydantic==1.8.2
PyYAML==3.12
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.3
uvicorn==0.11.8
